src/requirements.txt
pyyaml~=6.0.1
requests~=2.31.0
fastapi~=0.110.1
kafka-python
dagster
uvicorn